add some functions in run(), now it looks work

+++ alg.cpp
@@ -5,23 +5,23 @@
 #include <math.h>
 
 struct Pair{
-	long u;
-	long v;
+	int u;
+	int v;
 };
 
 struct y_element{
-	long u;
-	long d;
+	int u;
+	int d;
 };
 typedef y_element x_element;
 struct y_invert_list{
 	vector<y_element> l;
-	map<long,long> m;
+	map<int,int> m;
 };
 
 struct entry{
-	long u;
-	long d;
+	int u;
+	int d;
 	void *pointer;
 };
 
@@ -29,55 +29,55 @@ struct invert{
 	vector<entry> l;
 };
 
-map<long,invert> global_invert;
-map<long,map<long,long>> global_pair;
+map<int,invert> global_invert;
+map<int,map<int,int>> global_pair;
 
 typedef y_invert_list x_invert_list;
 
 struct in_element{
-	long x;
-	long d;
+	int x;
+	int d;
 };
 
 struct out_elemnt{
-	long y;
-	long d;
+	int y;
+	int d;
 };
 
 
 struct answer_entry{
-	vector<long> e;
+	vector<int> e;
 };
 
 //node u_k is an id, w_k is a word
-vector<long> getInverted(string key)// this function is equal to Pi in the draft
+vector<int> getInverted(string key)// this function is equal to Pi in the draft
 //input is the keyword, output is the list of nodes that contain this keyword
 {
 	
 }
 
-vector<in_element> IN(long u, long d)
+vector<in_element> IN(int u, int d)
 {// this function is equal to IN_d(u_k) in kewang's draft
 
 }
 
-vector<out_element> OUT(long u, long d)
+vector<out_element> OUT(int u, int d)
 {
 
 }
 
-map<long,y_element> constructY(string key,long D)
+map<int,y_element> constructY(string key,int D)
 {
-	map<long,y_element> Y;
-	vector<long> key_invert_list = getInverted(key);
-	for(long i=0; i<key_invert_list.size(); i++)
+	map<int,y_element> Y;
+	vector<int> key_invert_list = getInverted(key);
+	for(int i=0; i<key_invert_list.size(); i++)
 	{
-		long u = key_invert_list[i];
+		int u = key_invert_list[i];
 		vector<in_element> in_u = IN(u,D);
-		for(long j=0; j<in_u.size(); j++)
+		for(int j=0; j<in_u.size(); j++)
 		{
-			long x = in_u[j].x;
-			long d = in_u[j].d;
+			int x = in_u[j].x;
+			int d = in_u[j].d;
 			if(Y.find(x) == Y.end())
 			{
 				y_invert_list y;
@@ -86,7 +86,7 @@ map<long,y_element> constructY(string key,long D)
 			}
 			else
 			{
-				map<long,long> m = Y[x].m;
+				map<int,int> m = Y[x].m;
 				if(m.find(u)==m.end())
 				{
 					m[u] = d;
@@ -106,26 +106,26 @@ map<long,y_element> constructY(string key,long D)
 	return Y;
 }
 
-long nearest_ancestor(long y)
+int nearest_ancestor(int y)
 {
-	long x;
+	int x;
 	//use the T-structure to find the ancester
 	return x;// x is ancester of y;
 }
 
-long d_tree(long x, long y)
+int d_tree(int x, int y)
 {
 	return 2;
 }
 
-bool check(vector<long> a, long v, long D)
+bool check(vector<int> a, int v, int D)
 {
-	for(long i=0; i<a.size(); i++)
+	for(int i=0; i<a.size(); i++)
 	{
-		long u = a[i];
-		map<long,map<long,long>> ::iterator mmit;
+		int u = a[i];
+		map<int,map<int,int>> ::iterator mmit;
 		mmit = global_pair.find(u);
-		map<long,long> :: iterator mit;
+		map<int,int> :: iterator mit;
 		if(mmit != global_pair.end())
 		{
 			mit = (*mmit).find(v);
@@ -146,19 +146,19 @@ bool check(vector<long> a, long v, long D)
 	return true;
 }
 
-void step2(string key, long D, map<long,long> nodeAnswer, vector<answer_entry> answer )
+void step2(string key, int D, map<int,int> nodeAnswer, vector<answer_entry> answer )
 {
-	map<long,y_invert_list> Y = constructY( key, D);
-	map<long,x_invert_list> X;
-	for(map<long,y_invert_list> it=Y.begin(); it!=Y.end(); it++)
+	map<int,y_invert_list> Y = constructY( key, D);
+	map<int,x_invert_list> X;
+	for(map<int,y_invert_list> it=Y.begin(); it!=Y.end(); it++)
 	{
-		for(long i=0; i<(*it).second.l.size(); i++)
+		for(int i=0; i<(*it).second.l.size(); i++)
 		{
-			long y = (*it).first;
-			long u = (*it).second.l[i].u;
-			long d = (*it).second.l[i].d;
-			long x = nearest_ancestor(y,nodeAnswer);
-			long dT =  d_tree(x,y);
+			int y = (*it).first;
+			int u = (*it).second.l[i].u;
+			int d = (*it).second.l[i].d;
+			int x = nearest_ancestor(y,nodeAnswer);
+			int dT =  d_tree(x,y);
 			if( ( d + dT ) <= D )
 			{
 				if(X.find(x) == X.end())
@@ -169,7 +169,7 @@ void step2(string key, long D, map<long,long> nodeAnswer, vector<answer_entry> a
 				}
 				else 
 				{
-					map<long,long> m = X[x].m;
+					map<int,int> m = X[x].m;
 					if(m.find(u) == m.end())
 					{
 						m[u] = d;
@@ -186,24 +186,24 @@ void step2(string key, long D, map<long,long> nodeAnswer, vector<answer_entry> a
 		
 	}
 	// use a map to store all the d()
-	for(map<long,x_invert_list>:: iterator it = X.begin(); it!=X.end(); it++)
+	for(map<int,x_invert_list>:: iterator it = X.begin(); it!=X.end(); it++)
 	{
-		long x = (*it).first;
+		int x = (*it).first;
 		invert x_x_l = global_invert[x];
 		x_invert_list x_y_l = (*it).second;
-		for(long i=0; i<x_x_l.size(); i++)
+		for(int i=0; i<x_x_l.size(); i++)
 		{
-			long uj = x_x_l[i].u;
-			long dj = x_x_l[i].d;
-			for(long j=0; j<x_y_l.size(); j++)
+			int uj = x_x_l[i].u;
+			int dj = x_x_l[i].d;
+			for(int j=0; j<x_y_l.size(); j++)
 			{
-				long uk = x_y_l[j].u;
-				long dk = x_y_l[j].d;
+				int uk = x_y_l[j].u;
+				int dk = x_y_l[j].d;
 				if( (uj + uk) <D)
 				{
-					map<long,map<long,long>> ::iterator mmit;
+					map<int,map<int,int>> ::iterator mmit;
 					mmit = global_pair.find(uj);
-					map<long,long> :: iterator mit;
+					map<int,int> :: iterator mit;
 					if(mmit != global_pair.end())
 					{
 						mit = (*mmit).find(uk);
@@ -215,7 +215,7 @@ void step2(string key, long D, map<long,long> nodeAnswer, vector<answer_entry> a
 					}
 					else// uj doesn't exist in the map
 					{
-						map<long,long> mmap;
+						map<int,int> mmap;
 						mmap[uk] = dj + dk;
 						global_pair[uj] = mmap;
 					}
@@ -229,10 +229,10 @@ void step2(string key, long D, map<long,long> nodeAnswer, vector<answer_entry> a
 		}
 	}
 	vector<answer_entry> new_answer;
-	vector<long> key_list = getInverted( key);
-	for(long i=0; i<answer.size(); i++)
+	vector<int> key_list = getInverted( key);
+	for(int i=0; i<answer.size(); i++)
 	{
-		for(long j=0; j<key_list.size(); j++)
+		for(int j=0; j<key_list.size(); j++)
 		{
 			if(check(answer[i],key_list[j],D))
 			{
